drawer moved to a new component "UIDrawer:
