--- src/components/drawer/models/drawerVariant.ts
@@ -1,4 +1,0 @@
-/**
- * MUI Drawer types, "permanent", "persistent", "temporary"
- */
-export type DrawerVariant = "permanent" | "persistent" | "temporary" | undefined;
