+++ src/components/drawer/models/uidrawerVariant.ts
@@ -1,0 +1,4 @@
+/**
+ * MUI Drawer types, "permanent", "persistent", "temporary"
+ */
+export type UIDrawerVariant = "permanent" | "persistent" | "temporary" | undefined;
